# Improving performance of YAML storage