@renatomefi 2018-06-04
Hey @moufmouf that's super cool, thanks!

Since I didn't write this storage I'm wondering if there's a reason for the `char` iteration strategy, would you mind to read a bit a of the history in that file just to have a clue? I'll also check it myself soon!