@moufmouf 2018-06-04
Hey @renatomefi , thanks for the feedback.

Looking at the history, the "char" iteration strategy was there since the first commit (on Feb 23 2013):

https://github.com/php-vcr/php-vcr/blame/master/src/VCR/Storage/Yaml.php#L80

Commit done by @adri (ping!)

Interestingly, most of the code of the Yaml loader comes from the Json loader that was committed 3 days before (Feb 20 2013):

https://github.com/php-vcr/php-vcr/commit/aa5809778804ab04a31adbed6328a6a58e5850e2

The whole idea of those 2 classes is to avoid loading the whole JSON or YAML file at once, but to load it record by record instead (to avoid consuming too much memory).

As you can see in the docblock:

> This storage can be iterated while keeping the memory consumption to the
> amount of memory used by the largest record.

This was done from the very beginning this way (with `fgetc`). My guess is this is a "premature optimization" and was done so to be able to cope with cassettes of several hundreds of megabytes in size in the future without blowing the PHP memory limit. But I'm fairly confident this was never really tested on such a big cassette (because it would take several minutes to iterate a cassette that is 100MB long with the current code).