@adri 2018-06-04
Nice find! 
The problem I ran into when parsing the whole cassette was indeed blowing the memory_limit. The ideas was to limit the memory usage to the size of one request/response pair. Reading the whole line instead of each char makes a lot of sense.